requirements.txt
# ============================================================================
# Knowledge Nexus - Python Dependencies
# ============================================================================
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Markdown Processing
markdown>=3.5.0
python-frontmatter>=1.1.0
pymdown-extensions>=10.7.0
pygments>=2.17.0

# Search Engine
whoosh>=2.7.4

# Graph Processing
networkx>=3.2

# Configuration
pyyaml>=6.0.1

# Utilities
python-slugify>=8.0.0
httpx>=0.26.0
watchfiles>=0.21.0
rich>=13.7.0
aiofiles>=23.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
